@@ -66,22 +146,177 @@ include = [
     "fair_forge/artifacts/**/*",
     "README.md",
     "LICENSE",
+    "CONTRIBUTING.md",
+]
+exclude = [
+    "**/__pycache__",
+    "**/*.pyc",
+    "**/*.pyo",
+    "**/.DS_Store",
+    "**/test_*.py",
+    "tests/**",
 ]
 
-[tool.black]
-line-length = 120
-target-version = ['py310', 'py311', 'py312', 'py313']
-
+# Ruff configuration - Modern, fast Python linter & formatter
 [tool.ruff]
 line-length = 120
-target-version = "py310"
+target-version = "py313"
+extend-exclude = [
+    ".git",
+    ".venv",
+    "__pycache__",
+    "build",
+    "dist",
+    "*.egg-info",
+]
 
 [tool.ruff.lint]
-select = ["E", "F", "I", "N", "W"]
-ignore = ["E501"]
+# Enable comprehensive rule sets
+select = [
+    "E",      # pycodestyle errors
+    "F",      # pyflakes
+    "I",      # isort
+    "N",      # pep8-naming
+    "W",      # pycodestyle warnings
+    "UP",     # pyupgrade
+    "B",      # flake8-bugbear
+    "C4",     # flake8-comprehensions
+    "DTZ",    # flake8-datetimez
+    "T10",    # flake8-debugger
+    "EM",     # flake8-errmsg
+    "ISC",    # flake8-implicit-str-concat
+    "ICN",    # flake8-import-conventions
+    "PIE",    # flake8-pie
+    "PT",     # flake8-pytest-style
+    "RET",    # flake8-return
+    "SIM",    # flake8-simplify
+    "TID",    # flake8-tidy-imports
+    "TCH",    # flake8-type-checking
+    "ARG",    # flake8-unused-arguments
+    "PTH",    # flake8-use-pathlib
+    "PD",     # pandas-vet
+    "PL",     # pylint
+    "TRY",    # tryceratops
+    "RUF",    # ruff-specific rules
+]
 
+ignore = [
+    "E501",     # line too long (handled by formatter)
+    "PLR0913",  # too many arguments
+    "PLR0912",  # too many branches
+    "PLR0915",  # too many statements
+    "TRY003",   # avoid specifying long messages outside exception class
+    "PD901",    # avoid using 'df' as variable name
+    "B008",     # do not perform function calls in argument defaults (needed for FastAPI/Pydantic)
+]
+
+[tool.ruff.lint.per-file-ignores]
+"__init__.py" = ["F401", "E402"]  # Allow unused imports and module-level imports
+"tests/**/*.py" = ["ARG", "PLR2004", "S101"]  # Allow unused arguments, magic values, and asserts in tests
+
+[tool.ruff.lint.isort]
+known-first-party = ["fair_forge"]
+section-order = ["future", "standard-library", "third-party", "first-party", "local-folder"]
+
+[tool.ruff.format]
+quote-style = "double"
+indent-style = "space"
+skip-magic-trailing-comma = false
+line-ending = "auto"
+
+# MyPy configuration - Static type checking
 [tool.mypy]
-python_version = "3.10"
+python_version = "3.13"
 warn_return_any = true
 warn_unused_configs = true
-disallow_untyped_defs = false
+warn_redundant_casts = true
+warn_unused_ignores = true
+warn_no_return = true
+warn_unreachable = true
+strict_equality = true
+strict_concatenate = true
+check_untyped_defs = true
+disallow_untyped_defs = false  # Set to true when ready for full type coverage
+disallow_any_generics = false
+disallow_subclassing_any = false
+no_implicit_optional = true
+show_error_codes = true
+show_column_numbers = true
+pretty = true
+
+# Per-module options
+[[tool.mypy.overrides]]
+module = [
+    "nltk.*",
+    "transformers.*",
+    "sentence_transformers.*",
+    "hdbscan.*",
+    "umap.*",
+    "lakefs.*",
+    "boto3.*",
+]
+ignore_missing_imports = true
+
+# Pytest configuration
+[tool.pytest.ini_options]
+minversion = "8.0"
+testpaths = ["tests"]
+python_files = ["test_*.py", "*_test.py"]
+python_classes = ["Test*"]
+python_functions = ["test_*"]
+addopts = [
+    "-v",
+    "-ra",                          # Show summary of all test outcomes
+    "--strict-markers",
+    "--strict-config",
+    "--cov=fair_forge",
+    "--cov-branch",                 # Enable branch coverage
+    "--cov-report=term-missing:skip-covered",
+    "--cov-report=html",
+    "--cov-report=xml",
+    "--cov-fail-under=50",          # Fail if coverage below 50%
+    "--tb=short",                   # Shorter traceback format
+    "--maxfail=5",                  # Stop after 5 failures
+]
+markers = [
+    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
+    "integration: marks tests as integration tests",
+    "unit: marks tests as unit tests",
+    "requires_gpu: marks tests that require GPU",
+    "requires_api_key: marks tests that require API keys",
+]
+filterwarnings = [
+    "ignore::DeprecationWarning",
+    "ignore::PendingDeprecationWarning",
+    "ignore::UserWarning:pkg_resources",
+    "ignore::SyntaxWarning",  
+    "ignore::ImportWarning:umap",  
+    "error",  
+]
+
+# Coverage configuration
+[tool.coverage.run]
+source = ["fair_forge"]
+omit = [
+    "*/tests/*",
+    "*/__pycache__/*",
+    "*/venv/*",
+    "*/.venv/*",
+    "*/utils/logging.py",
+]
+branch = true
+
+[tool.coverage.report]
+exclude_lines = [
+    "pragma: no cover",
+    "def __repr__",
+    "raise AssertionError",
+    "raise NotImplementedError",
+    "if __name__ == .__main__.:",
+    "if TYPE_CHECKING:",
+    "if typing.TYPE_CHECKING:",
+    "@abstractmethod",
+    "@abc.abstractmethod",
+]
+precision = 2
+show_missing = true
