@@ -1,7 +1,5 @@
-from fair_forge import Guardian
-from fair_forge.schemas import ProtectedAttribute
-from fair_forge.schemas import GuardianBias
-from fair_forge.schemas import GuardianLLMConfig
+from fair_forge.core import Guardian
+from fair_forge.schemas import ProtectedAttribute, GuardianBias, GuardianLLMConfig
 from functools import partial
 from typing import Optional
 from transformers import AutoTokenizer
